+++ wxc_sdk/__init__.py
@@ -43,7 +43,7 @@ from dataclasses import dataclass
 
 __all__ = ['WebexSimpleApi']
 
-__version__ = '1.22.0'
+__version__ = '1.22.1'
 
 log = logging.getLogger(__name__)